Refactor anthology.py into subfolder
