+++ .gitignore
@@ -90,3 +90,4 @@ export/mods/*.xml
 # Even though we'll bring our own version of bibutils, we don't need it
 # inside the repository
 lib/tasks/bibutils_5.9
+**/__pycache__
