+++ bin/anthology/__init__.py
@@ -1,0 +1,4 @@
+from .anthology import Anthology
+from .people import PersonName, PersonIndex
+from .papers import Paper
+from .volumes import Volume
